refactor: rename/remove files related to ContactsTable.tsx
