--- src/components/ContactsLayout/_components/ContactsTable/Table-styles.less
@@ -1,114 +1,0 @@
-// By default, antd's Table is not taking its container height with a simple height: 100%,
-// so I had to customize the style.
-// Source: https://github.com/ant-design/ant-design/issues/23974#issuecomment-698787941
-.contacts-table {
-	width: 100%;
-	max-width: 100%;
-	height: 100%;
-
-	.ant-spin-nested-loading {
-		height: 100%;
-
-		.ant-spin-container {
-			height: 100%;
-			display: flex;
-			flex-flow: column nowrap;
-
-			.ant-table {
-				flex: auto;
-				overflow: hidden;
-
-				.ant-table-container {
-					height: 100%;
-					display: flex;
-					flex-flow: column nowrap;
-
-					.ant-table-header {
-						flex: none;
-					}
-
-					.ant-table-body {
-						flex: auto;
-						overflow: auto;
-						scrollbar-width: none;
-					}
-				}
-
-				.ant-table-pagination {
-					flex: none;
-				}
-			}
-		}
-	}
-
-	thead.ant-table-thead {
-		.contacts-table__header__light {
-			th {
-				background: #fffbf8;
-			}
-		}
-
-		.contacts-table__header__dark {
-			th {
-				background: var(--ant-color-primary-bg);
-			}
-		}
-	}
-
-	tbody.ant-table-tbody {
-		.favorite-button {
-			display: flex;
-			align-items: center;
-			justify-content: center;
-			color: var(--ant-color-primary);
-		}
-
-		.ant-table-expanded-row {
-			.ant-space {
-				margin-left: 2rem;
-			}
-		}
-
-		tr.ant-table-row {
-			&.even-row {
-				&__light td.ant-table-cell.ant-table-cell-row-hover {
-					background: #fdf0e3;
-				}
-
-				&__dark td.ant-table-cell.ant-table-cell-row-hover {
-					background: var(--ant-color-primary-bg-hover);
-				}
-			}
-
-			&.odd-row {
-				&__light td.ant-table-cell {
-					background: #fffbf8;
-
-					&.ant-table-cell-row-hover {
-						background: #fdf0e3;
-					}
-				}
-
-				&__dark td.ant-table-cell {
-					background: var(--ant-color-primary-bg);
-
-					&.ant-table-cell-row-hover {
-						background: var(--ant-color-primary-bg-hover);
-					}
-				}
-			}
-		}
-	}
-}
-
-.custom-filter-dropdown {
-	padding: 6px;
-
-	&.search {
-		width: 275px;
-	}
-
-	&__btns-container {
-		float: right;
-	}
-}
